requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
kafka-python>=2.0.2
pika>=1.2.0
python-dotenv>=0.19.0